feat: #89 SELECT文ビルダーを独立させた
